Refactor: Move the server entry and its modules under src/server
