+++ ai/skills/authoring/ssr-hydration.md
@@ -141,7 +141,7 @@ A server-rendered component passes through five phases from definition to intera
 
 ### renderToString
 
-`packages/component/src/render-to-string.js`
+`packages/component/src/server/render-to-string.js`
 
 The entry point for SSR. Takes a component class (returned by `defineComponent` with a `tagName`) and an attributes object, returns a complete DSD HTML string.
 
@@ -225,7 +225,7 @@ For `{#async}`, the server always renders `loadingContent` (never awaits the pro
 
 ### expandCustomElements
 
-`packages/component/src/expand-custom-elements.js`
+`packages/component/src/server/expand-custom-elements.js`
 
 After the top-level component renders, its HTML may contain nested custom element tags (e.g., `<ui-icon icon="star">` inside a button template). This function:
 
@@ -582,11 +582,11 @@ TESTING / BENCHING TRAPS (silent failures)
 
 ```
 packages/component/src/
-├── server.js                     @semantic-ui/component/server, the server functions and the server renderer's registration
-├── render-to-string.js           renderToString, the DSD render
-├── render-to-static-markup.js    renderToStaticMarkup, the markup alone
-├── server-template.js            createServerTemplate, the server instance both renders start from
-├── expand-custom-elements.js     Recursive custom element expansion
+├── server/index.js               @semantic-ui/component/server, the server functions and the server renderer's registration
+├── server/render-to-string.js    renderToString, the DSD render
+├── server/render-to-static-markup.js   renderToStaticMarkup, the markup alone
+├── server/server-template.js     createServerTemplate, the server instance both renders start from
+├── server/expand-custom-elements.js    Recursive custom element expansion
 ├── component-helpers.js          Shared: resolveAttributeAliases, getUIClasses
 ├── engines/native/base.js        WebComponentBase — hydrate(), canHydrate(), fullRender()
 └── engines/native/factory.js     Creates web component class with observedAttributes
