+++ packages/component/package.json
@@ -8,7 +8,7 @@
   "type": "module",
   "sideEffects": [
     "./src/index.js",
-    "./src/server.js",
+    "./src/server/index.js",
     "./src/engines/*/register.js"
   ],
   "main": "src/index.js",
@@ -24,8 +24,8 @@
     "./bare": "./src/bare.js",
     "./lit": "./src/engines/lit/register.js",
     "./server": {
-      "types": "./types/server.d.ts",
-      "import": "./src/server.js"
+      "types": "./types/server/index.d.ts",
+      "import": "./src/server/index.js"
     }
   },
   "scripts": {
