+++ packages/component/src/server/render-to-static-markup.js
