+++ packages/component/types/server/expand-custom-elements.d.ts
